Capstone Project/requirements.txt
chromadb==0.5.4
dask[dataframe]==2024.2.1
llama-cpp-python==0.2.80
numpy==1.26.4
pandas==2.2.2
rank-bm25==0.2.2
rouge-score==0.1.2
sentence-transformers==2.7.0
tqdm==4.66.5
